server/requirements.txt
PyPDF2
openai
flask
flask-cors
pikepdf
python-pptx
